+++ addOns/ascanrulesAlpha/CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Fixed
 - Fix an exception in Spring Actuator Information Leak scan rule when scanning responses without Content-Type header.
+- Correct path composition in Web Cache Deception scan rule.
 
 ## [38] - 2022-04-08
 ### Added